requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
dnspython==1.15.0
pkg-resources==0.0.0
py3dns==3.2.0
validate-email==1.3
